% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confirm_edges.R
\name{confirm_edges}
\alias{confirm_edges}
\title{Confirm Edges}
\usage{
confirm_edges(object, Rnew, method, alpha)
}
\arguments{
\item{object}{An object of classo \code{ggmncv}}

\item{Rnew}{Matrix. A correlation matrix of dimensions \emph{p} by \emph{p}.}

\item{method}{Character string. A correction method for multiple comparison (defaults to \code{fdr}).
Can be abbreviated. See \link[stats]{p.adjust}.}

\item{alpha}{Numeric. Significance level (defaults to \code{0.05}).}
}
\value{
An object of class \code{ggmncv}
}
\description{
Confirmatory testing of edges detected with data-driven
             model selection in an independent dataset.
}
\examples{
Y <- na.omit(bfi[,1:25])

Y_explore <- Y[1:1000,]

Y_confirm <- Y[1001:nrow(Y),]

fit <- ggmncv(cor(Y_explore),
              n = nrow(Y_explore))

confirm <- confirm_edges(fit,
                         Rnew = cor(Y_confirm),
                         method = "fdr",
                         alpha = 0.05)
}
