% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eip.R
\name{head.eip}
\alias{head.eip}
\title{Print the Head of \code{eip} Objects}
\usage{
\method{head}{eip}(x, n = 5, ...)
}
\arguments{
\item{x}{An object of class \code{eip}}

\item{n}{Numeric. Number of rows to print.}

\item{...}{Currently ignored.}
}
\value{
Invisibly returns a data frame containing the first \code{n} rows of
  \code{x$eip_results}. The returned data frame contains the leading edge
  relations and their edge inclusion probabilities, and is printed as a side
  effect.
}
\description{
Print the Head of \code{eip} Objects
}
