% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lw_estimator.R
\name{ledoit_wolf}
\alias{ledoit_wolf}
\title{Ledoit and Wolf Shrinkage Estimator}
\usage{
ledoit_wolf(Y, ...)
}
\arguments{
\item{Y}{A data matrix (or data.frame) of dimensions \emph{n} by \emph{p}.}

\item{...}{Currently ignored.}
}
\value{
Inverse correlation matrix.
}
\description{
Compute the Ledoit and Wolf shrinkage estimator of
the covariance matrix \insertCite{ledoit2004well}{GGMncv},
which can be used for the \code{initial} inverse covariance matrix
in \code{\link{ggmncv}}.
}
\examples{

# ptsd
Y <- ptsd[,1:5]

# shrinkage
ledoit_wolf(Y)

# non-reg
solve(cor(Y))
}
\references{
\insertAllCited{}
}
