% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eip.R
\name{plot.eip}
\alias{plot.eip}
\title{Plot Edge Inclusion 'Probabilities'}
\usage{
\method{plot}{eip}(x, color = "black", size = 1, ...)
}
\arguments{
\item{x}{An object of class \code{eip}}

\item{color}{Character string. Color for \code{geom_point}.}

\item{size}{Numeric. Size of \code{geom_point}.}

\item{...}{Currently ignored.}
}
\value{
An object of class \code{ggplot}
}
\description{
Plot Edge Inclusion 'Probabilities'
}
\examples{
\donttest{

# data
Y <- GGMncv::ptsd[,1:10]

# compute eip's
boot_samps <- boot_eip(Y, B = 10, progress = FALSE)


plot(boot_samps)

}
}
