% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivative.R
\name{plot.penalty_derivative}
\alias{plot.penalty_derivative}
\title{Plot \code{penalty_derivative} Objects}
\usage{
\method{plot}{penalty_derivative}(x, size = 1, ...)
}
\arguments{
\item{x}{An object of class \code{\link{penalty_derivative}}.}

\item{size}{Numeric. Line size in \code{geom_line}.}

\item{...}{Currently ignored.}
}
\value{
An object of class \code{ggplot}
}
\description{
Plot \code{penalty_derivative} Objects
}
\examples{
\donttest{
pen_deriv <- penalty_derivative(theta =  seq(-5,5,length.out = 10000),
                            lambda = 1,
                            gamma = c(0.01, 0.05, 0.1))
plot(pen_deriv)
}
}
