% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pen_function.R
\name{plot.penalty_function}
\alias{plot.penalty_function}
\title{Plot \code{penalty_function} Objects}
\usage{
\method{plot}{penalty_function}(x, size = 1, ...)
}
\arguments{
\item{x}{An object of class\code{\link{penalty_function}}.}

\item{size}{Numeric. Line size in \code{geom_line}.}

\item{...}{Currently ignored.}
}
\value{
An object of class \code{ggplot}
}
\description{
Plot \code{penalty_function} Objects
}
\examples{
\donttest{
func <- penalty_function(theta =  seq(-5,5,length.out = 10000),
                            lambda = 1,
                            gamma = c(0.01, 0.05, 0.1))
plot(func)
}
}
