% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eip.R
\name{print.eip}
\alias{print.eip}
\title{Print \code{eip} Objects}
\usage{
\method{print}{eip}(x, ...)
}
\arguments{
\item{x}{An object of class \code{eip}}

\item{...}{Currently ignored.}
}
\value{
Invisibly returns \code{x}, the input object of class \code{eip}.
  The method is called for its side effect of printing a data frame with
  the edge relations and their edge inclusion probabilities.
}
\description{
Print \code{eip} Objects
}
