% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Sachs}
\alias{Sachs}
\title{Data: Sachs Network}
\format{
A data frame containing 7466 cells (n = 7466) and flow cytometry
 measurements of 11 (p = 11) phosphorylated proteins and phospholipids
 \insertCite{sachs2002bayesian}{GGMncv}
}
\usage{
data("Sachs")
}
\description{
Protein expression in human immune system cells
}
\examples{
data("Sachs")

}
\references{
\insertAllCited{}
}
\keyword{datasets}
