% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.ggmncv.R
\name{coef.ggmncv}
\alias{coef.ggmncv}
\title{Regression Coefficients from \code{ggmncv} Objects}
\usage{
\method{coef}{ggmncv}(object, ...)
}
\arguments{
\item{object}{An Object of class \code{ggmncv}.}

\item{...}{Currently ignored.}
}
\value{
A matrix of regression coefficients.
}
\description{
Regression Coefficients from \code{ggmncv} Objects
}
\note{
The matrix of coefficients can be accessed via \code{coefs[1,]},
which provides the estimates for predicting the first node.

Further, the estimates are essentially computed with both
the outcome and predictors scaled to have mean 0 and
standard deviation 1.
}
\examples{

\donttest{

# data
Y <- GGMncv::ptsd[,1:5]

# correlations
S <- cor(Y)

# fit model
fit <- ggmncv(R = S, n = nrow(Y), progress = FALSE)

# regression
coefs <- coef(fit)

coefs

}

}
