% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmncv.R
\name{GGMncv}
\alias{GGMncv}
\title{GGMncv}
\usage{
GGMncv(
  Sigma,
  n,
  penalty = "selo",
  nonreg = TRUE,
  LLA = FALSE,
  steps = NULL,
  threshold = FALSE
)
}
\arguments{
\item{Sigma}{Correlation matrix}

\item{n}{Numeric. Sample size.}

\item{penalty}{Character string. Which penalty should be used (defaults to \code{selo}).}

\item{nonreg}{Logical. Should the precision matrix be reestimated, given the adjacency matrix
(defaults to TRUE)? This provides the \strong{nonregularized}, maximum likelhood estimate,
with constraints.}

\item{LLA}{Logical. Should the local linear approximation be used for maxmizing the penalized likelihood ?
The default is \code{FALSE} which results in a one-step approach (see details).}

\item{steps}{Numeric. How many steps if \code{LLA = TRUE} ? Set to \code{Inf} for iterating until
the algorithm converges.}

\item{threshold}{Logical. Should a threshold be applied (\code{1/sqrt(n-p-5)} ) ? Note that this
may be useful for small sample sizes to improve specificity (i.e., lower false positive rate),
but not needed otherwise (see details).}
}
\value{
An object of class \code{ggmncv}, including:
}
\description{
GGMncv
}
