% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmncv.R
\name{GGMncv}
\alias{GGMncv}
\title{GGMncv}
\usage{
GGMncv(
  x,
  n,
  lambda = NULL,
  n_lambda = 100,
  gamma = NULL,
  select = FALSE,
  L0_learn = FALSE,
  penalty = "atan",
  refit = FALSE,
  LLA = FALSE,
  method = "pearson",
  progress = TRUE,
  store = FALSE,
  ...
)
}
\arguments{
\item{x}{There are 2 options: either a \code{n} by \code{p} data matrix or a
\code{p} by \code{p} correlation matrix.}

\item{n}{Numeric. Sample size.}

\item{lambda}{Numeric. Tuning parameter governing the degress of penalization. Defaults to
\code{NULL} which results in fixing lambda to \code{sqrt(log(p)/n)}.}

\item{n_lambda}{Numeric. The number of regularization/thresholding parameters
(defaults to \code{100}).}

\item{gamma}{Numeric. Hyperparameter for the penalty function. Defaults to \code{0.1}
which is the recommended value for the default penalty (see details).}

\item{select}{Logical. Should lambda be selected with BIC (defaults to \code{FALSE})?}

\item{L0_learn}{Logical. Should lambda be selected based on the non-regularized
precision matrix (defaults to \code{FALSE}; see details).}

\item{penalty}{Character string. Which penalty should be used (defaults to \code{atan}).}

\item{refit}{Logical. Should the precision matrix be refitted, given the adjacency matrix
(defaults to \code{FALSE})? When set to \code{TRUE}, this provides the \strong{nonregularized},
maximum likelhood estimate with constraints.}

\item{LLA}{Logical. Should the local linear approximation be used for maxmizing the penalized likelihood ?
The default is \code{FALSE} (see details).}

\item{method}{Character string. Which correlation coefficient should be computed.
One of "pearson" (default), "spearman", or  "polychoric".}

\item{progress}{Logical. Should a progress bar be included (defaults to \code{TRUE}) ? Note that
this only applies when \code{select = TRUE}.}

\item{store}{Logical. Should all of the fitted models be saved (defaults to \code{NULL}). Note
this only applies when \code{select = TRUE}. and ignored otherwise
(the one model is saved.)}

\item{...}{Currently ignored.}
}
\value{
An object of class \code{ggmncv}, including:

\itemize{
\item \code{Theta} Inverse covariance matrix
\item \code{Sigma} Covariance matrix
\item \code{P} Weigthed adjacency matrix
\item \code{adj} Adjacency matrix
\item \code{lambda} Tuning parameter (i.e., sqrt(log(p)/n))
\item \code{fit} glasso fitted model (a list)
}
}
\description{
Estimate Gaussian graphical models with non-convex penalties.
}
\examples{
\donttest{
# data
Y <- GGMncv::ptsd

S <- cor(Y)

# fit model
fit <- GGMncv(S, n = nrow(Y))
}
}
