% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmncv.R
\name{GGMncv}
\alias{GGMncv}
\title{GGMncv}
\usage{
GGMncv(
  x,
  n,
  lambda = NULL,
  gamma = 0.1,
  penalty = "atan",
  nonreg = FALSE,
  LLA = FALSE,
  method = "pearson",
  vip = FALSE,
  vip_iter = 1000,
  progress = TRUE,
  ...
)
}
\arguments{
\item{x}{There are 2 options: either a \code{n} by \code{p} data matrix or a
\code{p} by \code{p} correlation matrix.}

\item{n}{Numeric. Sample size.}

\item{lambda}{Numeric. Tuning parameter governing the degress of penalization. Defaults to
\code{NULL} which results in fixing lambda to \code{sqrt(log(p)/n)}.}

\item{gamma}{Numeric. Hyperparameter for the penalty function. Defaults to \code{0.1}
which is the recommended value for the default penalty (see details).}

\item{penalty}{Character string. Which penalty should be used (defaults to \code{atan}).}

\item{nonreg}{Logical. Should the precision matrix be reestimated, given the adjacency matrix
(defaults to \code{FALSE})? When set to \code{TRUE}, this provides the \strong{nonregularized},
maximum likelhood estimate with constraints.}

\item{LLA}{Logical. Should the local linear approximation be used for maxmizing the penalized likelihood ?
The default is \code{FALSE} (see details).}

\item{method}{Character string. Which correlation coefficient should be computed.
One of "pearson" (default), "spearman", or  "polychoric".}

\item{vip}{Logical. Should the respective variable inclusion 'probabilities' be
computed (defaults to \code{FALSE})? This is accomplished with a
non-parametric bootstrap (see details)}

\item{vip_iter}{Numeric. How many number of bootstrap samples (defaults to 1000)?}

\item{progress}{Logical. Should a progress bar be included (defaults to \code{TRUE}) ? Note that
this only applies when \code{vip = TRUE}.}

\item{...}{Currently ignored.}
}
\value{
An object of class \code{ggmncv}, including:

\itemize{
\item \code{Theta} Inverse covariance matrix
\item \code{Sigma} Covariance matrix
\item \code{P} Weigthed adjacency matrix
\item \code{adj} Adjacency matrix
\item \code{lambda} Tuning parameter (i.e., sqrt(log(p)/n))
\item \code{fit} glasso fitted model (a list)
}
}
\description{
Estimate Gaussian graphical models with non-convex penalties.
}
\examples{
\donttest{
# data
Y <- GGMncv::ptsd

S <- cor(Y)

# fit model
fit <- GGMncv(S, n = nrow(Y))
}
}
