% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nct.R
\name{nct}
\alias{nct}
\title{Network Comparison Test}
\usage{
nct(
  Y_g1,
  Y_g2,
  iter = 1000,
  desparsify = TRUE,
  method = "pearson",
  FUN = NULL,
  cores = 1,
  progress = TRUE,
  ...
)
}
\arguments{
\item{Y_g1}{A matrix (or data.frame) of dimensions \emph{n} by \emph{p},
corresponding to the first dataset (\emph{p} must be the same
for \code{Y_g1} and \code{Y_g2}).}

\item{Y_g2}{A matrix of dimensions \emph{n} by \emph{p}, corresponding to the
second dataset (\emph{p} must be the same for \code{Y_g1} and \code{Y_g2}).}

\item{iter}{Numeric. Number of (Monte Carlo) permutations (defaults to \code{1000}).}

\item{desparsify}{Logical. Should the de-sparsified glasso estimator be
computed (defaults to \code{TRUE})? This is much faster,
as the tuning parameter is fixed to
\mjseqn{\lambda = \sqrt{log(p)/n}}.}

\item{method}{character string. Which correlation coefficient (or covariance)
is to be computed. One of "pearson" (default), "kendall",
or "spearman".}

\item{FUN}{A function or list of functions (defaults to \code{NULL}),
specifying custom test-statistics. See \strong{Examples}.}

\item{cores}{Numeric. Number of cores to use when executing the permutations in
parallel (defaults to \code{1}).}

\item{progress}{Should a progress bar be included ?}

\item{...}{Additional arguments passed to \code{\link{ggmncv}}.}
}
\description{
A re-implementation and extension of the permutation based
network comparison test introduced in \insertCite{van2017comparing;textual}{GGMncv}.
Such extensions include scaling to networks with many nodes and the option to
use custom test-statistics.
}
\references{
\insertAllCited{}
}
