% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.ggmncv}
\alias{predict.ggmncv}
\title{Predict method for \code{ggmncv} Objects}
\usage{
\method{predict}{ggmncv}(object, newdata = NULL, scale = "training")
}
\arguments{
\item{object}{An object of class \code{ggmncv}}

\item{newdata}{An optional data frame in which to look for variables with which to predict.
If omitted, the fitted values are used.}

\item{scale}{Character string. How should the new data (or test data) be centered and scaled ? The
default is to use the means and standard deviations of the trainging data. If set to anything other than
\code{training}, the new data is standardized with the \code{scale} function in based \code{R}.}
}
\value{
A matrix of predicted values
}
\description{
Predicted values based on a \code{ggmncv} object
}
\examples{


}
