% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eip.R
\name{print.eip}
\alias{print.eip}
\title{Print \code{eip} Objects}
\usage{
\method{print}{eip}(x, ...)
}
\arguments{
\item{x}{An object of class \code{eip}}

\item{...}{Currently ignored.}
}
\description{
Print \code{eip} Objects
}
