% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmncv.R
\name{print.ggmncv}
\alias{print.ggmncv}
\title{Print \code{ggmncv} Objects}
\usage{
\method{print}{ggmncv}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ggmncv}}

\item{...}{Currently ignored}
}
\description{
Print \code{ggmncv} Objects
}
