% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nct.R
\name{print.nct}
\alias{print.nct}
\title{Print \code{nct} Objects}
\usage{
\method{print}{nct}(x, ...)
}
\arguments{
\item{x}{An object of class \code{nct}}

\item{...}{Currently ignored.}
}
\description{
Print \code{nct} Objects
}
